llama-py/requirements.txt
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
ollama==0.2.1
pymongo==4.8.0
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
